+++ book-service/src/main/java/bjit/ursa/bookservice/repository/BookRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BookRepository extends JpaRepository<BookEntity , Integer> {
+public interface BookRepository extends JpaRepository<BookEntity , Long> {
     boolean existsByBookNameAndAuthorName(String bookName, String genre);
     //Optional<BookEntity> findByBookNameAndGenre(String bookName, String genre);
 }
